+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Erl2ex is currently pre-alpha software. Expect significant backwards-incompatibl
 ## v0.0.6 (not yet released)
 
 *   Convert string() and nonempty_string() Erlang types to the preferred Elixir equivalents, to avoid an Elixir warning.
+*   Support for environment variable interpolation in include paths.
 
 ## v0.0.5 (2016-01-11)